rounding errors with bigger durations.

+++ vpro-shared-elasticsearch-client/src/main/java/nl/vpro/elasticsearchclient/ElasticSearchIterator.java
@@ -229,7 +229,7 @@ public class ElasticSearchIterator<T>  implements CountedIterator<T> {
     public Optional<Instant> getETA() {
         if (getCount() != null && getCount() != 0 && getTotalSize().isPresent()) {
             Duration duration = Duration.between(start, Instant.now());
-            Duration estimatedTotalDuration = Duration.ofNanos(duration.toNanos() * getTotalSize().get() / getCount());
+            Duration estimatedTotalDuration = Duration.ofNanos((long) (duration.toNanos() * (getTotalSize().get().doubleValue() / getCount())));
             return Optional.of(start.plus(estimatedTotalDuration));
         } else {
             return Optional.empty();
